Fix django db errors
