+++ bf_opencore/tests/test_asset.py
@@ -14,23 +14,26 @@ from freezegun import freeze_time
 from bf_opencore import models
 
 
+@pytest.mark.django_db
 def test_get_empty_assets(auth_client):
     """Test that asset list is empty unless we do something special."""
     assets = auth_client.get("/api/assets/")
     assert assets.data["count"] == 0
 
 
+@pytest.mark.django_db
 def test_create_asset(auth_client):
     """Creating an empty asset makes it available via the API."""
-    dummy_asset = models.Asset.objects.create()
+    _ = models.Asset.objects.create()
     assets = auth_client.get("/api/assets/")
     assert assets.status_code == 200
     assert assets.data["count"] == 1
 
 
+@pytest.mark.django_db
 def test_get_asset_csv(auth_client):
     """We can specify CSV format."""
-    dummy_asset = models.Asset.objects.create()
+    _ = models.Asset.objects.create()
     # import pdb ; pdb.set_trace()
     response = auth_client.get("/api/assets/", HTTP_ACCEPT="text/csv")
     assert response.status_code == 200
@@ -38,9 +41,10 @@ def test_get_asset_csv(auth_client):
     assert len(response.content.splitlines()) == 2  # header row + 1 asset
 
 
+@pytest.mark.django_db
 def test_get_asset_csv_specify_fields(auth_client):
     """With CSV format we generally would specify which headers we want."""
-    dummy_asset = models.Asset.objects.create(name="spam", ip_address="10.0.0.1")
+    _ = models.Asset.objects.create(name="spam", ip_address="10.0.0.1")
     # import pdb ; pdb.set_trace()
     response = auth_client.get(
         "/api/assets/?fields=name,ip_address,model", HTTP_ACCEPT="text/csv"
@@ -53,9 +57,10 @@ def test_get_asset_csv_specify_fields(auth_client):
     assert asset == ["spam", "10.0.0.1", ""]
 
 
+@pytest.mark.django_db
 def test_export_assets_json(auth_client):
     """We can export assets to JSON."""
-    dummy_asset = models.Asset.objects.create()
+    _ = models.Asset.objects.create()
     # import pdb ; pdb.set_trace()
     response = auth_client.get("/api/assets/", HTTP_ACCEPT="application/json")
     assert response.status_code == 200
@@ -63,6 +68,7 @@ def test_export_assets_json(auth_client):
     assert json.loads(response.content)
 
 
+@pytest.mark.django_db
 def test_api_create_asset(asset_edit_client):
     """Create an asset with authorized client."""
     client = asset_edit_client
@@ -78,6 +84,7 @@ def test_api_create_asset(asset_edit_client):
     assert asset["hostname"] == "nospam"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_maconly(asset_edit_client):
     """Create an asset with authorized client."""
     client = asset_edit_client
@@ -93,6 +100,7 @@ def test_api_create_asset_maconly(asset_edit_client):
     assert asset["mac_address"] == "00:00:00:00:00:01"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_addinventory_maconly(asset_edit_client):
     """Create an asset, replicating 'add inventory.
 
@@ -113,6 +121,7 @@ def test_api_create_asset_addinventory_maconly(asset_edit_client):
     assert asset["ip_address"] is None
 
 
+@pytest.mark.django_db
 def test_api_create_asset_addinventory_empty_mac(asset_edit_client):
     """Create an asset, replicating 'add inventory.
 
@@ -133,6 +142,7 @@ def test_api_create_asset_addinventory_empty_mac(asset_edit_client):
     assert asset["ip_address"] is None
 
 
+@pytest.mark.django_db
 def test_api_create_asset_addinventory_empty_mac_times_two(asset_edit_client):
     """Create an asset, replicating 'add inventory'.
 
@@ -161,6 +171,7 @@ def test_api_create_asset_addinventory_empty_mac_times_two(asset_edit_client):
     assert assets.data["count"] == 2
 
 
+@pytest.mark.django_db
 def test_api_create_asset_unauthorized(auth_client):
     """Can't create an asset with an unauthorized client."""
     client = auth_client
@@ -175,6 +186,7 @@ def test_api_create_asset_unauthorized(auth_client):
     assert len(assets.data["results"]) == 0
 
 
+@pytest.mark.django_db
 def test_api_create_get_asset(auth_client, asset_edit_client):
     """Create an asset, read with less-authorized client."""
     response = asset_edit_client.post(
@@ -189,6 +201,7 @@ def test_api_create_get_asset(auth_client, asset_edit_client):
     assert asset["hostname"] == "nospam"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_open_ports(asset_edit_client):
     """Create an asset with authorized client.
 
@@ -206,6 +219,7 @@ def test_api_create_asset_open_ports(asset_edit_client):
     assert asset["open_ports_tcp"] == [80, 443, 8000]
 
 
+@pytest.mark.django_db
 def test_api_create_patch_asset(auth_client, asset_edit_client):
     """Create an asset, then patch."""
     response = asset_edit_client.post(
@@ -229,6 +243,7 @@ def test_api_create_patch_asset(auth_client, asset_edit_client):
     assert asset["hostname"] == "spam"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_displayname(asset_edit_client):
     """Create an asset with a display_name."""
     client = asset_edit_client
@@ -242,6 +257,7 @@ def test_api_create_asset_displayname(asset_edit_client):
     assert response.data["display_name"] == "foobar"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_displayname_name(asset_edit_client):
     """Create an asset with a display_name *and* a name.
 
@@ -262,6 +278,7 @@ def test_api_create_asset_displayname_name(asset_edit_client):
     assert response.data["display_name"] == "foobar"
 
 
+@pytest.mark.django_db
 def test_api_create_asset_displayname_name_2(asset_edit_client):
     """Create an asset with a display_name *and* a name.
 
@@ -279,6 +296,7 @@ def test_api_create_asset_displayname_name_2(asset_edit_client):
     assert response.data["display_name"] == "foobar"
 
 
+@pytest.mark.django_db
 def test_api_update_asset_displayname(asset_edit_client):
     """Create an asset, update display_name later."""
     client = asset_edit_client
@@ -298,6 +316,7 @@ def test_api_update_asset_displayname(asset_edit_client):
     assert response.data["display_name"] == "spam"
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(
     raises=AssertionError,
     reason="Not sure why, but we *are* allowed to patch.  "
@@ -333,12 +352,13 @@ def test_api_create_unauth_patch_asset(auth_client, asset_edit_client):
     assert asset["hostname"] == "nospam"
 
 
+@pytest.mark.django_db
 def test_unauth_patch_asset(auth_client):
     """Create an asset, patch with less-authorized client.
 
     (auth_client is 'authenticated', not 'authorized')
     """
-    dummy_asset = models.Asset.objects.create()
+    _ = models.Asset.objects.create()
     response = auth_client.get("/api/assets/")
     assert response.status_code == 200  # 200 = Created
     assert response.json()["count"] == 1
@@ -357,6 +377,7 @@ def test_unauth_patch_asset(auth_client):
     assert asset["hostname"] is None  # Still None
 
 
+@pytest.mark.django_db
 def test_create_many_assets(auth_client):
     """Creating multiple assets makes them available via the API."""
     asset_names = ["foo", "bar", "baz", "xyzzy", "spam", "ham", "eggs"]
@@ -366,6 +387,7 @@ def test_create_many_assets(auth_client):
     assert assets.data["count"] == len(asset_names)
 
 
+@pytest.mark.django_db
 def test_retrieve_one_asset(auth_client):
     """Creating an asset makes it available via the API."""
     asset_names = ["foo", "bar", "baz", "xyzzy", "spam", "ham", "eggs"]
@@ -376,6 +398,7 @@ def test_retrieve_one_asset(auth_client):
     assert len(assets.data["results"]) == 1
 
 
+@pytest.mark.django_db
 def test_one_asset_details(auth_client):
     """Creating an asset makes its details available via the API."""
     asset_names = ["foo", "bar", "baz", "xyzzy", "spam", "ham", "eggs"]
@@ -387,6 +410,7 @@ def test_one_asset_details(auth_client):
     assert asset["hostname"] == "xyzzy"
 
 
+@pytest.mark.django_db
 def test_patch_asset(asset_edit_client):
     """Patching an asset field updates the asset in the database."""
     client = asset_edit_client
@@ -405,6 +429,7 @@ def test_patch_asset(asset_edit_client):
     assert spam_asset.hostname == "nospam"
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_string(asset_edit_client):
     """Patching ports with a string of integers.
 
@@ -432,6 +457,7 @@ def test_patch_asset_open_ports_tcp_string(asset_edit_client):
     assert spam_asset.open_ports_tcp == [80, 443, 8000]
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_list(asset_edit_client):
     """Patching ports with a string of integers.
 
@@ -458,6 +484,7 @@ def test_patch_asset_open_ports_tcp_list(asset_edit_client):
     assert spam_asset.open_ports_tcp == [80, 443, 8000]
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_list_bad(asset_edit_client):
     """Patching ports with a string of integers.
 
@@ -479,6 +506,7 @@ def test_patch_asset_open_ports_tcp_list_bad(asset_edit_client):
     assert response.status_code == 400
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_bad(asset_edit_client):
     """Patching ports with a string of integers... but they are bad."""
     client = asset_edit_client
@@ -501,6 +529,7 @@ def test_patch_asset_open_ports_tcp_bad(asset_edit_client):
     assert spam_asset.open_ports_tcp == []
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_null(asset_edit_client):
     """Patching ports with Null sets the port list to empty.
 
@@ -526,6 +555,7 @@ def test_patch_asset_open_ports_tcp_null(asset_edit_client):
     assert spam_asset.open_ports_tcp == []
 
 
+@pytest.mark.django_db
 def test_patch_asset_open_ports_tcp_empty(asset_edit_client):
     """Patching with empty string sets the port list to empty.
 
@@ -551,6 +581,7 @@ def test_patch_asset_open_ports_tcp_empty(asset_edit_client):
     assert spam_asset.open_ports_tcp == []
 
 
+@pytest.mark.django_db
 def test_set_name_empty(asset_edit_client):
     """PATCHing a hostname to an empty string works."""
     client = asset_edit_client
@@ -569,6 +600,7 @@ def test_set_name_empty(asset_edit_client):
     assert spam_asset.hostname == ""
 
 
+@pytest.mark.django_db
 def test_set_name_null(asset_edit_client):
     """PATCHing a hostname to `null` works."""
     client = asset_edit_client
@@ -588,6 +620,7 @@ def test_set_name_null(asset_edit_client):
     assert spam_asset.hostname is None
 
 
+@pytest.mark.django_db
 def test_field_histogram(auth_client):
     """Field histogram works with one field."""
     models.Asset.objects.create(manufacturer="Bar")
@@ -606,18 +639,21 @@ def test_field_histogram(auth_client):
     assert [x["count"] for x in qset] == [3, 2, 1]
 
 
+@pytest.mark.django_db
 def test_field_histogram_no_field(auth_client):
     """Fail to specify field=<Asset field name> specified: HTTP 400"""
     response = auth_client.get("/api/assets/histogram/")
     assert response.status_code == 400
 
 
+@pytest.mark.django_db
 def test_field_histogram_no_such_field(auth_client):
     """Specify field=<nonsense>: HTTP 400"""
     response = auth_client.get("/api/assets/histogram/", {"field": "asdf"})
     assert response.status_code == 400
 
 
+@pytest.mark.django_db
 def test_api_duplicate_ips(auth_client):
     """Check for duplicate IP addresses in the asset population."""
     response = auth_client.get("/api/assets/duplicate_ips/")
@@ -638,6 +674,7 @@ def test_api_duplicate_ips(auth_client):
     ]
 
 
+@pytest.mark.django_db
 def test_api_duplicate_ips_one(auth_client):
     """Check for duplicate IP addresses in the asset population."""
     response = auth_client.get("/api/assets/duplicate_ips/")
@@ -657,6 +694,7 @@ def test_api_duplicate_ips_one(auth_client):
     ]
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(
     raises=AssertionError,
     reason="not allowed to spell out 'exact' in URL for some reason.",
@@ -680,6 +718,7 @@ def test_api_duplicate_ips_one_spell_exact(auth_client):
     ]
 
 
+@pytest.mark.django_db
 def test_api_duplicate_ips_one_prefix_robust(auth_client):
     """Check for duplicate IP addresses in the asset population."""
     response = auth_client.get("/api/assets/duplicate_ips/")
@@ -701,6 +740,7 @@ def test_api_duplicate_ips_one_prefix_robust(auth_client):
     ]
 
 
+@pytest.mark.django_db
 def test_fetch_by_os(auth_client):
     """Assets can be fetched by OS field."""
     a1 = models.Asset.objects.create(os="Windows XP")
@@ -728,6 +768,7 @@ def test_fetch_by_os(auth_client):
     assert ids == set([a1.id, a2.id])
 
 
+@pytest.mark.django_db
 def test_app_sw_version_needs_update(auth_client):
     """Test whether assets need software updates."""
     oldest = models.Asset.objects.create(
@@ -756,6 +797,7 @@ def test_app_sw_version_needs_update(auth_client):
     assert response.json()["needs_update"] is False
 
 
+@pytest.mark.django_db
 def test_nonsense_app_sw_version_needs_update(auth_client):
     """Test whether a silly asset needs an update."""
     a = models.Asset.objects.create(
@@ -772,6 +814,7 @@ def test_nonsense_app_sw_version_needs_update(auth_client):
     assert resp.json()["needs_update"] is True
 
 
+@pytest.mark.django_db
 def test_app_sw_version_not_needs_update(auth_client):
     """Test whether two equal assets need updates."""
     a1 = models.Asset.objects.create(
@@ -790,6 +833,7 @@ def test_app_sw_version_not_needs_update(auth_client):
     assert response2.json()["needs_update"] is False
 
 
+@pytest.mark.django_db
 def test_app_sw_no_version_needs_update(auth_client):
     """Test whether assets without software versions need updates."""
     asset = models.Asset.objects.create(manufacturer="Foo", model="Bar")
@@ -800,6 +844,7 @@ def test_app_sw_no_version_needs_update(auth_client):
     assert response.json()["needs_update"] is False
 
 
+@pytest.mark.django_db
 def test_api_create_asset_mac_autofill_nic(asset_edit_client):
     """Create an asset with NIC vendor."""
     client = asset_edit_client
@@ -816,6 +861,7 @@ def test_api_create_asset_mac_autofill_nic(asset_edit_client):
     assert asset["nic_vendor"] == "Apple, Inc."
 
 
+@pytest.mark.django_db
 def test_api_create_asset_mac_reject_nic(asset_edit_client):
     """Provided NIC vendor will be silently ignored."""
     client = asset_edit_client
@@ -837,6 +883,7 @@ def test_api_create_asset_mac_reject_nic(asset_edit_client):
     assert asset["nic_vendor"] == "Apple, Inc."
 
 
+@pytest.mark.django_db
 def test_upsert_create(asset_edit_client):
     """Create a new asset via upsert endpoint."""
     macaddr = "00:03:b1:b5:b6:48"
@@ -857,6 +904,7 @@ def test_upsert_create(asset_edit_client):
     assert asset.mac_address == macaddr
 
 
+@pytest.mark.django_db
 def test_upsert_update(asset_edit_client):
     """Update an existing asset via upsert endpoint."""
     # Create existing asset in database
@@ -880,6 +928,7 @@ def test_upsert_update(asset_edit_client):
     assert str(asset.ip_address) == "10.0.0.1"
 
 
+@pytest.mark.django_db
 def test_upsert_no_mac_address(asset_edit_client):
     """Upsert endpoint ignores calls that lack a MAC address."""
     response = asset_edit_client.post(
@@ -895,6 +944,7 @@ def test_upsert_no_mac_address(asset_edit_client):
     assert models.Asset.objects.count() == 0
 
 
+@pytest.mark.django_db
 def test_upsert_no_mac_address_duplicate(asset_edit_client):
     """Call upsert endpoint twice, with the same IP address.
 
@@ -925,6 +975,7 @@ def test_upsert_no_mac_address_duplicate(asset_edit_client):
     assert models.Asset.objects.count() == 1
 
 
+@pytest.mark.django_db
 def test_upsert_ipv6(asset_edit_client):
     """Call upsert endpoint with an ipv6 address.
 
@@ -945,6 +996,7 @@ def test_upsert_ipv6(asset_edit_client):
     assert models.Asset.objects.count() == 1
 
 
+@pytest.mark.django_db
 def test_upsert_many_fields(asset_edit_client):
     """Call upsert endpoint with a fields typically provided by sniffer."""
     response = asset_edit_client.post(
@@ -968,6 +1020,7 @@ def test_upsert_many_fields(asset_edit_client):
     assert models.Asset.objects.count() == 1
 
 
+@pytest.mark.django_db
 def test_upsert_ipv4(asset_edit_client):
     """Call upsert endpoint with a "ipv4_address" field.
 
@@ -990,6 +1043,7 @@ def test_upsert_ipv4(asset_edit_client):
     assert str(asset.ip_address) == "10.0.0.1"
 
 
+@pytest.mark.django_db
 def test_upsert_bad_key(asset_edit_client):
     """Call upsert endpoint with a bad key in the JSON."""
     with pytest.raises(django.core.exceptions.FieldDoesNotExist):
@@ -1005,6 +1059,7 @@ def test_upsert_bad_key(asset_edit_client):
         )
 
 
+@pytest.mark.django_db
 def test_upsert_open_port_tcp(asset_edit_client):
     """Call upsert endpoint with a "open_port_tcp" field.
 
@@ -1027,6 +1082,7 @@ def test_upsert_open_port_tcp(asset_edit_client):
     assert asset.open_ports_tcp == [80]
 
 
+@pytest.mark.django_db
 def test_upsert_identifier(asset_edit_client):
     """Call upsert endpoint with a "identifier" field.
 
@@ -1048,6 +1104,7 @@ def test_upsert_identifier(asset_edit_client):
     assert asset.name == "Alaris 8100"
 
 
+@pytest.mark.django_db
 @pytest.mark.parametrize(
     "date_range, num_assets",
     [
@@ -1097,6 +1154,7 @@ def test_asset_date_range(date_range, num_assets, auth_client):
         assert res.data["count"] == num_assets
 
 
+@pytest.mark.django_db
 def test_external_key_non_connector(db, auth_client):
     """Test that external_links/ detail route doesn't barf on non-connector
     external key.
