+++ bf_opencore/tests/test_asset_relations.py
@@ -4,6 +4,7 @@ Uses built-in pytest-django text fixtures from
 http://pytest-django.readthedocs.io/en/latest/helpers.html
 """
 
+import pytest
 import django.utils.timezone
 
 from bf_opencore import models
@@ -20,6 +21,7 @@ from bf_opencore import models
 # Tags
 
 
+@pytest.mark.django_db
 def test_get_asset_tags_obsolete(auth_client):
     """Test old /api/assets/<n>/tags way to get tags associated with asset.
 
@@ -40,6 +42,7 @@ def test_get_asset_tags_obsolete(auth_client):
     assert {t["id"] for t in tags} == {tag_red.pk, tag_green.pk}
 
 
+@pytest.mark.django_db
 def test_get_asset_tags_new(auth_client):
     """Test new /api/tags/?asset=<n> way to get tags associated with asset."""
     asset = models.Asset.objects.create(hostname="foo.com")
@@ -54,6 +57,7 @@ def test_get_asset_tags_new(auth_client):
     assert {t["id"] for t in tags} == {tag_red.pk, tag_green.pk}
 
 
+@pytest.mark.django_db
 def test_get_asset_vulnerabilities_obsolete(auth_client, asset_vulnerabilities):
     """Test old /api/assets/<n>/vulnerabilities way to get vulns for asset."""
     asset = asset_vulnerabilities[0]
@@ -61,6 +65,7 @@ def test_get_asset_vulnerabilities_obsolete(auth_client, asset_vulnerabilities):
     assert response.status_code == 404
 
 
+@pytest.mark.django_db
 def test_get_asset_vulnerabilities_new(auth_client, asset_vulnerabilities):
     """Test new /api/vulnerabilities/?asset=<n> way to get vulns for asset."""
     (asset, vulnerability_red, vulnerability_green) = asset_vulnerabilities
@@ -73,6 +78,7 @@ def test_get_asset_vulnerabilities_new(auth_client, asset_vulnerabilities):
     }
 
 
+@pytest.mark.django_db
 def test_get_asset_asset_vulnerabilities_obsolete(auth_client, asset_vulnerabilities):
     """Test old /api/assets/<n>/assetvulnerabilities route for asset_vulns."""
     asset = asset_vulnerabilities[0]
@@ -80,6 +86,7 @@ def test_get_asset_asset_vulnerabilities_obsolete(auth_client, asset_vulnerabili
     assert response.status_code == 404
 
 
+@pytest.mark.django_db
 def test_get_asset_asset_vulnerabilities_new(auth_client, asset_vulnerabilities):
     """Test new /api/assetvulnerabilities/?asset=<n> route for asset_vulns."""
     (asset, vulnerability_red, vulnerability_green) = asset_vulnerabilities
@@ -94,99 +101,10 @@ def test_get_asset_asset_vulnerabilities_new(auth_client, asset_vulnerabilities)
 
 
 ################
-# Scans
-
-
-def test_get_asset_scans_obsolete(auth_client):
-    """Test old /api/assets/<n>/scans way to get scans of asset.
-
-    NOTE: will remove this route; then change assertion to
-          assert response.status_code == 404 (or 405)
-    """
-    raise NotImplementedError("Connectors have been removed")
-    connector = models.Connector.objects.create(id="spam")
-    c_task_1 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="1"
-    )
-    c_task_2 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="2"
-    )
-    c_task_3 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="3"
-    )
-    asset = models.Asset.objects.create(hostname="foo.com")
-    asset_dummy = models.Asset.objects.create(hostname="spam.com")
-    aux_fields = {
-        "date_scanned": django.utils.timezone.now(),
-        "num_vulnerabilities": 0,
-        "num_plugins": 0,
-    }
-    scan_red = models.Scan.objects.create(
-        asset=asset, connector_task=c_task_1, **aux_fields
-    )
-    scan_green = models.Scan.objects.create(
-        asset=asset, connector_task=c_task_2, **aux_fields
-    )
-    dummy_scan = models.Scan.objects.create(
-        asset=asset_dummy, connector_task=c_task_1, **aux_fields
-    )
-    dummy_scan = models.Scan.objects.create(
-        asset=asset_dummy, connector_task=c_task_3, **aux_fields
-    )
-    response = auth_client.get(f"/api/assets/{asset.pk}/scans/")
-    # assert response.status_code == 405
-    # assert response.status_text == "Method Not Allowed"
-    scans = response.data["results"]
-    assert len(scans) == 2
-    assert {t["id"] for t in scans} == {scan_red.pk, scan_green.pk}
-
-
-def test_get_asset_scans_new(auth_client):
-    """Test new /api/scans/?asset=<n> way to get scans of asset."""
-    raise NotImplementedError("Connectors have been removed")
-    connector = models.Connector.objects.create(id="spam")
-    c_task_1 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="1"
-    )
-    c_task_2 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="2"
-    )
-    c_task_3 = models.ConnectorTask.objects.create(
-        connector=connector, celery_task_id="3"
-    )
-    asset = models.Asset.objects.create(hostname="foo.com")
-    asset_dummy = models.Asset.objects.create(hostname="spam.com")
-    aux_fields = {
-        "date_scanned": django.utils.timezone.now(),
-        "num_vulnerabilities": 0,
-        "num_plugins": 0,
-    }
-    scan_red = models.Scan.objects.create(
-        asset=asset, connector_task=c_task_1, **aux_fields
-    )
-    scan_green = models.Scan.objects.create(
-        asset=asset, connector_task=c_task_2, **aux_fields
-    )
-    dummy_scan = models.Scan.objects.create(
-        asset=asset_dummy, connector_task=c_task_1, **aux_fields
-    )
-    dummy_scan = models.Scan.objects.create(
-        asset=asset_dummy, connector_task=c_task_3, **aux_fields
-    )
-    response = auth_client.get(f"/api/scans/?asset={asset.pk}")
-    scans = response.data["results"]
-    assert len(scans) == 2
-    assert {t["id"] for t in scans} == {scan_red.pk, scan_green.pk}
-    # Sanity check (should get all scans.)
-    response = auth_client.get("/api/scans/")
-    scans = response.data["results"]
-    assert len(scans) == 4
-
-
-################
 # Networks
 
 
+@pytest.mark.django_db
 def test_get_asset_network_old_api(admin_client):
     """Ensure we can determine which assets belong in network.
 
@@ -203,6 +121,7 @@ def test_get_asset_network_old_api(admin_client):
     assert networks[0]["id"] == network.pk
 
 
+@pytest.mark.django_db
 def test_get_asset_network_new_api(admin_client):
     """Ensure we can determine which assets belong in network."""
     asset = models.Asset.objects.create(ip_address="10.0.0.1")
@@ -217,3 +136,4 @@ def test_get_asset_network_new_api(admin_client):
     networks = response.data["results"]
     assert len(networks) == 2
     assert {n["id"] for n in networks} == {network_blue.pk, network_red.pk}
+
