+++ bf_opencore/tests/test_autocomplete_field.py
@@ -1,8 +1,10 @@
 """Test autocomplete interface."""
 
 from bf_opencore import models
+import pytest
 
 
+@pytest.mark.django_db
 def test_autocomplete_field(auth_client, complete_us):
     """Test that /autocomplete_field/ returns the right set."""
     assert models.Asset.objects.count() == 6
@@ -12,6 +14,7 @@ def test_autocomplete_field(auth_client, complete_us):
     assert set(manufs.json()) == set(["Foo", "Bar"])
 
 
+@pytest.mark.django_db
 def test_autocomplete_field_with_constraints(auth_client, complete_us):
     """Test that /autocomplete_field/ respects constraints."""
     assert models.Asset.objects.count() == 6
@@ -21,6 +24,7 @@ def test_autocomplete_field_with_constraints(auth_client, complete_us):
     assert set(manufs.json()) == set(["One", "Two", "Three"])
 
 
+@pytest.mark.django_db
 def test_autocomplete_field_empty(auth_client):
     """Test that /autocomplete_field/ returns empty when it should."""
     assert models.Asset.objects.count() == 0
