+++ bf_opencore/tests/test_csv.py
@@ -3,35 +3,11 @@
 import os
 import tempfile
 
-import django.core.management
 import pytest
 
 import bf_opencore
 import bf_opencore.celery
 from bf_opencore.models import Asset
-
-
-@pytest.fixture
-def setup_db(db):
-    """Create connector objects."""
-    raise NotImplementedError("Connectors have been removed")
-    # ConnectorTask objects are linked to Connector objects with a foreign key
-    # relationship.  We need those objects to be available.
-    django.core.management.call_command("create_connectors")
-
-    # First enable Connector that's disabled by default
-    connector = Connector.objects.get(id="csv")
-    connector.enabled = True
-    connector.save()
-
-
-@pytest.fixture
-def no_nwk_field():
-    """Configure the CSV connector to not require network field."""
-    raise NotImplementedError("Connectors have been removed")
-    connector = Connector.objects.get(id="csv")
-    connector.settings["require_network_info"] = False
-    connector.save()
 
 
 class TestCTX:
@@ -59,447 +35,3 @@ class TestCTX:
 
     ct = TestConnectorTask()
 
-
-def write_tempfile(text, bom_utf8=False):
-    """Write text (ostensibly, CSV) to a temp file and return the filename."""
-    csvfd, filename = tempfile.mkstemp(suffix=".csv")
-    if bom_utf8:
-        # The 'utf-8-sig' special encoding inserts the byte order mark
-        # '0xef, 0xbb, 0xdf' at the beginning of the file (or strips it
-        # off, if it's there when reading.)
-        # NOTE: in general it is *discouraged* to use this BOM.  We use
-        # it here only to test that we're robust against it if it is in
-        # a file we come across.
-        #
-        # https://docs.python.org/3/library/codecs.html#encodings-and-unicode
-        encoding = "utf-8-sig"
-    else:
-        encoding = None
-    with open(csvfd, "w", encoding=encoding) as fh:
-        fh.write(text)
-    return filename
-
-
-def test_csv_simple(setup_db):
-    """Run a connector task and verify that it was connected to the DB.
-    This will implicitely verify that the Celery signals correctly created
-    and associated with a ConnectorTask database object.  That means there will
-    be exactly one ConnectorTask object in the database.
-    """
-    raise NotImplementedError("Connectors have been removed")
-    # It would be really confusing to break up a CSV line on to multiple lines
-    # pylint: disable=line-too-long
-    field_mapping = {
-        "serial_number": "Serial",
-        "mac_address": "MAC",
-        "ip_address": "IP",
-        "manufacturer": "Manufacturer",
-        "model": "Model",
-    }
-    filename = write_tempfile(
-        "Manufacturer,Model,IP,MAC,Serial,other\n"
-        "Hospira,Plum A+,10.10.0.13,01:00:00:00:00:13,a013,This thing is totally broken\n"
-        "Alaris,8100,192.168.0.1,01:00:00:00:00:01,a001,Last serviced by Ben\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.filter(mac_address="01:00:00:00:00:13").exists()
-    assert Asset.objects.filter(mac_address="01:00:00:00:00:01").exists()
-
-
-def test_csv_non_default_field_mapping(setup_db):
-    """Run a connector task and verify that it was connected to the DB.
-    This will implicitely verify that the Celery signals correctly created
-    and associated with a ConnectorTask database object.  That means there will
-    be exactly one ConnectorTask object in the database.
-    """
-    raise NotImplementedError("Connectors have been removed")
-    # It would be really confusing to break up a CSV line on to multiple lines
-    # pylint: disable=line-too-long
-    field_mapping = {
-        "serial_number": "MySerial",
-        "mac_address": "MyMAC",
-        "ip_address": "MyIP",
-        "manufacturer": "MyManufacturer",
-        "model": "MyModel",
-    }
-    filename = write_tempfile(
-        "MyManufacturer,MyModel,MyIP,MyMAC,MySerial,Mother\n"
-        "Hospira,Plum A+,10.10.0.13,01:00:00:00:00:13,a013,This thing is totally broken\n"
-        "Alaris,8100,192.168.0.1,01:00:00:00:00:01,a001,Last serviced by Ben\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 2
-    assert Asset.objects.filter(mac_address="01:00:00:00:00:13").exists()
-    assert Asset.objects.filter(mac_address="01:00:00:00:00:01").exists()
-
-
-def test_no_ip_or_mac(setup_db):
-    """Do not import assets with neither IP nor MAC address."""
-    raise NotImplementedError("Connectors have been removed")
-    # It would be really confusing to break up a CSV line on to multiple lines
-    # pylint: disable=line-too-long
-    field_mapping = {
-        "mac_address": "MyMAC",
-        "ip_address": "MyIP",
-    }
-    filename = write_tempfile(
-        "MyIP,MyMAC\n,,\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 0
-
-
-def test_bad_mac(setup_db):
-    """Test that a weird MAC address precludes saving asset."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "external_keys__tms": "AssetNumber",
-        "mac_address": "MACAddress",
-    }
-    filename = write_tempfile(
-        "AssetNumber,MACAddress\n126,72.18.10.62\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 0
-
-
-def test_bad_ip(setup_db):
-    """Test that a bad IP address precludes saving asset."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "external_keys__tms": "AssetNumber",
-        "ip_address": "IP1",
-    }
-    filename = write_tempfile(
-        "AssetNumber,IP1\n126,192.168.0.345\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 0
-
-
-def test_to_asset_model_ok(setup_db):
-    """Mapping to Asset model fields works."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "manufacturer": ["Manoof"],
-        "model": "Model",
-        "ip_address": "LAN 1 IP",
-    }
-    filename = write_tempfile(
-        "Model,AssetNumber,LAN 1 IP,Manoof\nX4000,1,10.0.0.1,FooCorp\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    asset = Asset.objects.last()
-    assert asset.manufacturer == "FooCorp"
-    assert asset.model == "X4000"
-
-
-def test_to_asset_model_bad_invalidated1(setup_db):
-    """Field that doesn't exist in Asset unwelcome in field mapping."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "manufacturer": ["Manoof"],
-        "nonexistent_field": "Hi there",
-    }
-    filename = write_tempfile(
-        "AssetNumber,Manoof\n124,FooCorp\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Failed"
-    assert "'nonexistent_field' key in FieldMap" in ct.stderr
-
-
-def test_to_asset_model_bad_invalidated2(setup_db):
-    """Field that doesn't exist in Asset unwelcome in field mapping."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "manufacturer": ["M"],
-    }
-    filename = write_tempfile(
-        "AssetNumber,Manoof\n124,FooCorp\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Failed"
-    assert "keymap value 'M' does not map" in ct.stderr
-
-
-def test_fail_validation(setup_db):
-    """Bad field values should fail validation upon conversion."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "mac_address": "MACAddress",
-        "external_keys__tms": "AssetNumber",
-    }
-    filename = write_tempfile(
-        "AssetNumber,MACAddress\n1,0:1:2:3:4:5\n2,hamburger\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-
-    # Import the asset when when MAC address is valid
-    asset = Asset.objects.get(mac_address="0:1:2:3:4:5")
-    assert asset.external_keys["tms"] == "1"
-
-    # Don't import when the MAC is invalid
-    assert not Asset.objects.filter(external_keys__tms="2").exists()
-
-
-def test_sync_match_on_integer_key_id(setup_db):
-    """Looking up an asset by its TMS Key ID finds the right asset."""
-    raise NotImplementedError("Connectors have been removed")
-    Asset.objects.create(
-        manufacturer="Foo",
-        model="Bar",
-        external_keys={"tms": "123"},
-    )
-    field_mapping = {
-        "manufacturer": "Manufacturer",
-        "ip_address": "LAN 1 IP",
-        "external_keys__tms": "AssetNumber",
-    }
-    filename = write_tempfile(
-        "AssetNumber,LAN 1 IP,Manufacturer\n123,10.0.0.1,Different\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 1
-    asset = Asset.objects.last()
-    assert asset.manufacturer == "Different"
-
-
-def test_sync_match_on_string_key_id(setup_db):
-    """Looking up an asset by its (string) TMS PK finds the right asset."""
-    raise NotImplementedError("Connectors have been removed")
-    Asset.objects.create(
-        manufacturer="Foo",
-        model="Bar",
-        external_keys={"tms": "ONETWOTHREE"},
-    )
-    field_mapping = {
-        "external_keys__tms": "Asset #",
-        "manufacturer": "Manufacturer",
-        "ip_address": "LAN 1 IP",
-    }
-    filename = write_tempfile(
-        "Asset #,LAN 1 IP,Manufacturer\nONETWOTHREE,10.0.0.1,Different\n",
-    )
-    asset = Asset.objects.last()
-    assert asset.manufacturer == "Foo"
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert Asset.objects.count() == 1
-    asset = Asset.objects.last()
-    assert asset.manufacturer == "Different"
-
-
-def test_risk_score_simple(setup_db):
-    """Looking up an asset by its (string) TMS PK finds the right asset."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "mac_address": "MAC",
-        "asset_risk_factors__tms": "RISK",
-    }
-    filename = write_tempfile(
-        "MAC,RISK\n11:22:33:44:55:66,5.7\n",
-    )
-
-    # Run CSV connector
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-
-    # Check ConnectorTask database entry of recently run connector
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert ct.stderr == ""
-
-    # Check Asset created by connector
-    assert Asset.objects.count() == 1
-    asset = Asset.objects.last()
-
-    # Check AssetRiskFactor created by connector
-    assert asset.asset_risk_factors.count() == 1
-    arf = asset.get_risk_factor("tms")
-    assert arf.value == 5.7
-
-
-def test_risk_score_invalid(setup_db):
-    """Non-numeric string for risk score."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "mac_address": "MAC",
-        "asset_risk_factors__tms": "RISK",
-    }
-    filename = write_tempfile(
-        "MAC,RISK\n11:22:33:44:55:66,five point seven\n",
-    )
-
-    # Run CSV connector
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-
-    # Check ConnectorTask database entry of recently run connector
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert ct.stderr == ""
-
-    # Check Asset created by connector.  The asset is created, and an error
-    # occurs when adding a risk factor to the asset.
-    assert Asset.objects.count() == 1
-    asset = Asset.objects.last()
-
-    # Risk is coerced to 0
-    assert asset.asset_risk_factors.count() == 1
-    arf = asset.get_risk_factor("tms")
-    assert arf.value == 0.0
-
-
-def test_risk_score_none(setup_db):
-    """Field mappping fails to find a value for risk score."""
-    raise NotImplementedError("Connectors have been removed")
-    field_mapping = {
-        "mac_address": "MAC",
-        "asset_risk_factors__tms": "RISK",
-    }
-    filename = write_tempfile(
-        "MAC,RISK\n11:22:33:44:55:66,\n",  # RISK value is empty string
-    )
-
-    # Run CSV connector
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-
-    # Check ConnectorTask database entry of recently run connector
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert "" in ct.stderr
-
-    # Check Asset created by connector
-    assert Asset.objects.count() == 1
-    asset = Asset.objects.last()
-
-    # Risk is coerced to 0
-    assert asset.asset_risk_factors.count() == 1
-    arf = asset.get_risk_factor("tms")
-    assert arf.value == 0.0
-
-
-def test_category(setup_db):
-    """Test import to the category field through a field mapping."""
-    raise NotImplementedError("Connectors have been removed")
-    # It would be really confusing to break up a CSV line on to multiple lines
-    field_mapping = {
-        "mac_address": "MAC",
-        "category": "Category",
-    }
-    filename = write_tempfile(
-        "MAC,Category\n11:22:33:44:55:66,Infusion Pump\n",
-    )
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    asset = Asset.objects.get(mac_address="11:22:33:44:55:66")
-    assert asset.category == "Infusion Pump"
