+++ bf_opencore/tests/test_csv_custom_field.py
@@ -1,52 +1,16 @@
 """Test CSV integration."""
 
 import os
+import pytest
 
-import bf_opencore
 import bf_opencore.celery
 from bf_opencore.csv import process_csv
 from bf_opencore.models import Asset, AssetCustomField, AssetCustomFieldName
 
-from .test_csv import TestCTX, write_tempfile
+from .test_csv import TestCTX
+from bf_opencore.tests.utils import write_tempfile
 
-
-def test_csv_import_with_custom_field(setup_db, no_nwk_field):
-    """We can import from CSV into a custom field."""
-    raise NotImplementedError("Connectors have been removed")
-    Asset.objects.create(
-        manufacturer="Foo",
-        model="Bar",
-        external_keys={"other_cmms": "ONETWOTHREE"},
-    )
-    shininess_field_name = AssetCustomFieldName.objects.create(field_name="Shininess")
-    field_mapping = {
-        "external_keys__other_cmms": "Asset #",
-        "asset_custom_fields__Shininess": "How shiny it is",
-        "manufacturer": "Manufacturer",
-    }
-    filename = write_tempfile(
-        "Asset #,Manufacturer,How shiny it is\nONETWOTHREE,Different,Very shiny\n",
-    )
-    asset = Asset.objects.last()
-    asset_custom_fields = AssetCustomField.objects.filter(asset=asset)
-    assert asset_custom_fields.count() == 0
-    bf_opencore.csv.main.apply(
-        kwargs={
-            "filename": filename,
-            "field_mapping": field_mapping,
-        }
-    )
-    os.unlink(filename)
-    ct = ConnectorTask.objects.get()
-    assert ct.status == "Success"
-    assert asset_custom_fields.count() == 1
-
-    asset_custom_fields = AssetCustomField.objects.filter(asset=asset)
-    asset_shininess_field = asset_custom_fields.get(field=shininess_field_name)
-    asset_shininess = asset_shininess_field.value_text
-    assert asset_shininess == "Very shiny"
-
-
+@pytest.mark.django_db
 def test_process_csv_with_custom_field(setup_db):
     """We can import from CSV into a custom field.
 
@@ -88,6 +52,7 @@ def test_process_csv_with_custom_field(setup_db):
     assert asset_shininess == "Very shiny"
 
 
+@pytest.mark.django_db
 def test_process_csv_with_custom_field_underscores(setup_db):
     """We can import from CSV into a custom field that contains underscores."""
     Asset.objects.create(
@@ -129,6 +94,7 @@ def test_process_csv_with_custom_field_underscores(setup_db):
     assert asset_site_description == "Very shiny"
 
 
+@pytest.mark.django_db
 def test_process_csv_with_custom_field_spaces(setup_db):
     """We can import from CSV into a custom field that contains spaces."""
     Asset.objects.create(
@@ -168,3 +134,4 @@ def test_process_csv_with_custom_field_spaces(setup_db):
     )
     asset_site_description = asset_site_description_field.value_text
     assert asset_site_description == "Very shiny"
+
