+++ bf_opencore/tests/test_network.py
@@ -9,6 +9,7 @@ from bf_opencore import models
 # models do have 'objects' member, but it's being lazy loaded
 
 
+@pytest.mark.django_db
 def test_create_empty_network(nwk_authorized_client):
     """Creating a network without a name is not allowed."""
     response = nwk_authorized_client.post(
@@ -19,6 +20,7 @@ def test_create_empty_network(nwk_authorized_client):
     assert response.status_code == 400
 
 
+@pytest.mark.django_db
 def test_create_named_network(nwk_authorized_client):
     """Create a network without CIDR via the API.
 
@@ -36,6 +38,7 @@ def test_create_named_network(nwk_authorized_client):
     assert network.cidr == []
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(reason="Open-core has no role-based write permissions")
 def test_create_named_network_unauth(auth_client):
     """Don't create a network with unauthorized client."""
@@ -64,6 +67,7 @@ CIDR_TEST_DATA = [
 ]
 
 
+@pytest.mark.django_db
 @pytest.mark.parametrize("supplied_cidr, resulting_cidr", CIDR_TEST_DATA)
 def test_create_network_then_cidr(supplied_cidr, resulting_cidr, nwk_authorized_client):
     """Create a network then add one or more CIDR."""
@@ -83,6 +87,7 @@ def test_create_network_then_cidr(supplied_cidr, resulting_cidr, nwk_authorized_
     assert network.cidr == resulting_cidr
 
 
+@pytest.mark.django_db
 @pytest.mark.parametrize("supplied_cidr, resulting_cidr", CIDR_TEST_DATA)
 def test_get_cidr(supplied_cidr, resulting_cidr, nwk_authorized_client):
     """Create a network + CIDR, then get CIDR."""
@@ -101,6 +106,7 @@ def test_get_cidr(supplied_cidr, resulting_cidr, nwk_authorized_client):
     assert response.status_code == 200
 
 
+@pytest.mark.django_db
 def test_delete_cidr(nwk_authorized_client):
     """Create a network, add CIDR, delete CIDR."""
     response = nwk_authorized_client.post(
@@ -126,6 +132,7 @@ def test_delete_cidr(nwk_authorized_client):
     assert network.cidr == []
 
 
+@pytest.mark.django_db
 def test_change_cidr_one(nwk_authorized_client):
     """Create a network, add CIDR, change CIDR."""
     response = nwk_authorized_client.post(
@@ -147,6 +154,7 @@ def test_change_cidr_one(nwk_authorized_client):
     assert network.cidr == ["10.0.1.3/32"]
 
 
+@pytest.mark.django_db
 @pytest.mark.parametrize("test_case", range(1, len(CIDR_TEST_DATA)))
 def test_change_cidr(nwk_authorized_client, test_case):
     """Create a network, add CIDR, change CIDR.
@@ -181,6 +189,7 @@ def test_change_cidr(nwk_authorized_client, test_case):
     assert network.cidr == CIDR_TEST_DATA[test_case][1]
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(
     raises=ValueError,
     reason="Create Nwk w/CIDR is not yet implemented, fails "
@@ -203,6 +212,7 @@ def test_create_cidr_network(nwk_authorized_client):
     # assert network.cidr == ['10.0.1.2']
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(
     raises=AssertionError,
     reason="There's a bug in CIDR validation code.  Documented in Github issue #1443.",
@@ -224,6 +234,7 @@ def test_patch_bad_cidr(nwk_authorized_client):
     assert response.status_code == 400  # This should fail
 
 
+@pytest.mark.django_db
 def test_cidr_bad_json(nwk_authorized_client):
     """Verify list of CIDRs is valid JSON."""
     # Create a network
@@ -247,6 +258,7 @@ def test_cidr_bad_json(nwk_authorized_client):
 # Test Asset - Network connections
 
 
+@pytest.mark.django_db
 def test_asset_in_network_old_api(nwk_authorized_client):
     """Ensure we can determine network membership."""
     dummy_asset = models.Asset.objects.create(ip_address="10.0.0.1")
@@ -256,6 +268,7 @@ def test_asset_in_network_old_api(nwk_authorized_client):
     assert response.status_code == 404
 
 
+@pytest.mark.django_db
 def test_asset_in_network_new_api(nwk_authorized_client):
     """Ensure we can determine network membership."""
     asset = models.Asset.objects.create(ip_address="10.0.0.1")
@@ -268,6 +281,7 @@ def test_asset_in_network_new_api(nwk_authorized_client):
     assert assets[0]["id"] == asset.id
 
 
+@pytest.mark.django_db
 def test_asset_big_network(nwk_authorized_client):
     """Ensure these things work also in other networks."""
     asset_ips = [
@@ -300,3 +314,4 @@ def test_asset_big_network(nwk_authorized_client):
     assets = response_big.data["results"]
     assert len(assets) == 9
     assert {a["ip_address"] for a in assets} == set(asset_ips)
+
