+++ bf_opencore/tests/test_reverse.py
@@ -9,6 +9,7 @@ from bf_opencore import models
 # models do have 'objects' member, but it's being lazy loaded
 
 
+@pytest.mark.django_db
 def test_reverse_asset_tags():
     """Reverse bf_opencore:asset-tags == /api/asset/<i>/tags."""
     # import pdb ; pdb.set_trace()
@@ -16,6 +17,7 @@ def test_reverse_asset_tags():
     assert url == "/api/assets/1/tags/"
 
 
+@pytest.mark.django_db
 def test_reverse_group_single():
     """Reverse bf_opencore:asset-tags == /api/asset/<i>/tags."""
     # import pdb ; pdb.set_trace()
@@ -23,6 +25,7 @@ def test_reverse_group_single():
     assert url == "/api/groups/1/"
 
 
+@pytest.mark.django_db
 def test_reverse_group():
     """Reverse bf_opencore:asset-tags == /api/asset/<i>/tags."""
     # import pdb ; pdb.set_trace()
@@ -30,6 +33,7 @@ def test_reverse_group():
     assert url == "/api/groups/"
 
 
+@pytest.mark.django_db
 @pytest.mark.xfail(
     raises=django.urls.exceptions.NoReverseMatch,
     reason="Route /api/assets/1/groups/ doesn't exist (and shouldn't)",
@@ -41,6 +45,7 @@ def test_reverse_asset_groups():
     assert url == "/api/assets/1/groups/"
 
 
+@pytest.mark.django_db
 def test_asset_tags_api(auth_client):
     """API for tags associated with asset works."""
     asset_obj = models.Asset.objects.create()
@@ -49,6 +54,7 @@ def test_asset_tags_api(auth_client):
     assert res.json()["count"] == 0
 
 
+@pytest.mark.django_db
 def test_asset_groups_api(auth_client):
     """API for groups associated with asset does NOT work this way.
 
@@ -57,3 +63,4 @@ def test_asset_groups_api(auth_client):
     asset_obj = models.Asset.objects.create()
     res = auth_client.get(f"/api/assets/{asset_obj.id}/groups/")
     assert res.status_code == 404
+
